Separate parcel from Tully example, update docs to final paper
