+++ docs/bibliography.bib
@@ -439,14 +439,15 @@ doi  = {10.1039/C2CS35200A}
   doi = {10.1175/1520-0450(1980)019<1037:ANIOSF>2.0.CO;2}
 }
 
-@article{Tully2022,
+@article{Tully2023,
+  title = {Assessing predicted cirrus ice properties between two deterministic ice formation parameterizations},
   author = {Tully, C. and Neubauer, D. and Lohmann, U.},
-  title = {Technical Note: assessing predicted cirrus ice properties between two deterministic ice formation parameterizations},
-  journal = {EGUsphere},
-  volume = {2022},
-  year = {2022},
-  pages = {1--25},
-  doi = {10.5194/egusphere-2022-1057}
+  journal = {Geoscientific Model Development},
+  volume = {16},
+  year = {2023},
+  number = {10},
+  pages = {2957--2973},
+  doi = {10.5194/gmd-16-2957-2023}
 }
 
 @article{Vehkamaki2002,
